requirements.txt
elit-tokenizer==1.0
PySide6==6.6.3.1
PySide6_Addons==6.6.3.1
PySide6_Essentials==6.6.3.1
shiboken6==6.6.3.1
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.5.1
